Moved bootstrap-specific code out of generic footer
