+++ packages/fastui/src/components/footer.tsx
@@ -15,28 +15,17 @@ export const FooterComp = (props: FooterProps) => {
     link.mode = link.mode || 'footer'
     return link
   })
-  const linkProp = useClassName(props, { el: 'link' })
-  const extraSeparatorProp = useClassName(props, { el: 'extra-separator' })
-  const extraProp = useClassName(props, { el: 'extra' })
+  const extraTextClassName = useClassName(props, { el: 'extra' })
   return (
-    <div className={useClassName(props, { el: 'contents' })}>
-      <footer className={useClassName(props)}>
-        <ul className={useClassName(props, { el: 'link-list' })}>
-          {links.map((link, i) => (
-            <li className={linkProp}>
-              <LinkComp key={i} {...link} />
-            </li>
-          ))}
-        </ul>
-        {props.extraText && props.links.length > 0 && <div className={extraSeparatorProp} />}
-        {props.extraText && (
-          <>
-            <div className={extraProp}>
-              <TextComp {...props.extraText} />
-            </div>
-          </>
-        )}
-      </footer>
-    </div>
+    <footer className={useClassName(props)}>
+      {links.map((link, i) => (
+        <LinkComp key={i} {...link} />
+      ))}
+      {props.extraText && (
+        <div className={extraTextClassName}>
+          <TextComp {...props.extraText} />
+        </div>
+      )}
+    </footer>
   )
 }
